Moved the function to get information in Model layer
