+++ backend/src/music/config.json
@@ -85,7 +85,7 @@
     ]
   },
   {
-    "name": "THE WITHER",
+    "name": "THE WITCHER",
     "tag": {
       "example": "This contains the information on which the playlist interface should be based",
       "pic": "https://psv4.userapi.com/s/v1/d/fkW2xl-ZNlCnTvV4C1gSIR_yqxJpJU0Q0nEOkj0-EtRmz5_pOPgYs60Z8F8pgS1hwTa2UFFK3w95SdNhVf8QUzfyHW3oczILSfq-a3DL8mJHhnt4y8OoqQ/IMG_6797.png",
