+++ frontend/src/redux/reducers/audioSlice.ts
@@ -3,6 +3,7 @@ import axios from "axios";
 import { loadAudioState } from "../../utils/localStorage";
 import Bugsnag from "@bugsnag/js";
 import {API_URL} from "../../config";
+import { fetchPlaylistTracks } from "../../model/getMusics";
 
 interface Track {
     id: number;
@@ -30,40 +31,6 @@ const initialState: AudioState = {
     currentTime: savedState?.currentTime || 0,
     error: null,
 };
-
-export const fetchPlaylistTracks = createAsyncThunk(
-    "audio/fetchPlaylistTracks",
-    async (playlistId: number, { rejectWithValue })=> {
-        try {
-            const response = await axios.post(
-                API_URL + "api/music/musics",
-                { "id": playlistId },
-                { withCredentials: true }
-            );
-            if (response.data.error) {
-                throw new Error(response.data.error);
-            }
-
-            const tracks = response.data.musics.map((track: Track) => ({
-                id: track.id,
-                playlist_id: track.playlist_id,
-                name: track.name,
-                author: track.author,
-                filename: track.filename,
-                tag: track.tag,
-            }));
-
-            return tracks;
-        } catch (error: unknown) {
-            if (error instanceof Error) {
-                Bugsnag.notify(error.message)
-                return rejectWithValue(error.message);
-            }
-            Bugsnag.notify("Unknown error occurred")
-            return rejectWithValue("Unknown error occurred");
-        }
-    }
-)
 
 const audioSlice = createSlice({
     name: "audio",
